Change teleport_on_cursor to use the vanilla TeleportHelper
Delete the Command_Teleporter.cs file. The code also had a bug where it
expected the current body to have the KinematicCharacterMotor component,
which fails for flyers.

Now takes into account the character's foot position properly instead
of hardcoding an offset of 5, which used to result to a hop for human-
sized characters.