Modify StringFinder partial matching to list names in matching priority
The priority is (partial at the beginning of the word > partial anywhere)
and (partial matching a higher percentage of the word > lower). This makes
list commands behave exactly the same as the new autocompletion options.

The only exceptions are the dynamic director cards and interactables, as
the list commands use more aliases to match a partial than the respective
autocomplete variables. This only affects two commands and shouldn't have
an effect with many partial strings since it's only the top one that
matters.