Improved autocompletion (#177)
* Create an enhanced autocompletion implementation

This version uses a command's argument signature to derive autofill
options on a per-argument basis. It can also store "variables" which can
be reused in multiple commands. As such it suffices to define the list of
all item names, for example, in one single place.

* Transition commands to the new autocompletion implemention

Deletes the file Code/AutoCompletion.cs as it's obselete.

* Update hooks for the new autocomplete

* Update the filters for StringFinder.RemoveSpaceAndAlike

Ignore format items that appear in some language strings, e.g., for elites
Overloading {0}.

Remove round brackets from names as the characters are ignored in the
console input and separate arguments. For example, getting the suggestion
"DiosBestFriend(Consumed)" for give_item would effectively be the same as
`give_item DiosBestFriend Consumed`.

* Fix some command signature inconsistencies

list_skin, next_stage, perm_enable, perm_mod did not abide by the syntax
for autocomplete.

Correctly state that change_team, spawn_as, and respawn require the player
argument from server. Also fix a bug in their respective methods where
a server could get away by passing the default value for the player.

* Make the current command's signature appear while typing in the console

* Modify StringFinder partial matching to list names in matching priority

The priority is (partial at the beginning of the word > partial anywhere)
and (partial matching a higher percentage of the word > lower). This makes
list commands behave exactly the same as the new autocompletion options.

The only exceptions are the dynamic director cards and interactables, as
the list commands use more aliases to match a partial than the respective
autocomplete variables. This only affects two commands and shouldn't have
an effect with many partial strings since it's only the top one that
matters.

* Update README to refer to the autocompletion API

* Blacklist comma for RemoveSpacesAndAlike, move hyphen to the end

* Make tab and ctrl+tab to also cycle through the autocomplete options

* Fix the autocompletion dropdown not resetting when switching arguments

* Fix the give_equip arg signature to include "random" for autocomplete

* Improve the Autocompletion documentation

- Partial matching prioritises matches at the beggining of a string
- Explicitly describe all steps to include DebugToolkit in another mod
- Highlight type the code block for C#

* Make assembly operations safe and give user freedom for assembly input

* Add the new autocompletion attr to commands that came from merged PRs

---------

Co-authored-by: Quentin <xiaoxiao921@hotmail.fr>