A bunch of small fixes/improvements (kill_all, spawn_as, teleport_on_cursor) (#176)
* Fix kill_all "collection was modified" and incorrect kill count

If a new character is created the moment one is killed, e.g., Artifact
of Soul or Mending Elites, the team member enumerable is modified during
iteration.

* Change spawn_as to use master.TransformBody

* Change teleport_on_cursor to use the vanilla TeleportHelper

Delete the Command_Teleporter.cs file. The code also had a bug where it
expected the current body to have the KinematicCharacterMotor component,
which fails for flyers.

Now takes into account the character's foot position properly instead
of hardcoding an offset of 5, which used to result to a hop for human-
sized characters.

* Revert "Change teleport_on_cursor to use the vanilla TeleportHelper"

This reverts commit 1b215bc84945e0e8732c02c809e4a8a64228c961.

* Fix teleport_on_cursor null when KinematicCharacterMotor is missing

This applies the change of the reverted fix by using TeleportHelper, but
wrapped in the networked component which was deleted previously.