Make noclip toggleable without a body and change disconnect hook
If the server drops the session or kicks a client, it does not trigger the
Disconnect hook. OnStopClient is called in all cases.

Allowing toggling while dead also addresses the edge case of the client
leaving the session while in that state. Toggling shouldn't require a body,
as that is only necessary for modifying collisions which can be taken care
of when respawned.