Make the networked object DontDestroyOnLoad
With the game object persistent between scenes noclip doesn't have to
toggle off when changing stages.

The added event to disable noclip when a run ends makes it consistent with
disabling it when disonnecting. This also preserves the old functionality
where noclip would be toggled off after each stage. While within a run it
isn't as helpful, it can be so at the end of it.