Make the network object persistent between scene transitions and address noclip toggling cases (#182)
* Make noclip toggleable without a body and change disconnect hook

If the server drops the session or kicks a client, it does not trigger the
Disconnect hook. OnStopClient is called in all cases.

Allowing toggling while dead also addresses the edge case of the client
leaving the session while in that state. Toggling shouldn't require a body,
as that is only necessary for modifying collisions which can be taken care
of when respawned.

* Control whether noclip logs its toggling action

Useful for preventing logs when the state is toggled off and on again to
properly reset itself during Update.

Also remove the now unused method SetUseGravity that was left over from an
earlier commit: ac6fe9ec2889dc0c70b95dff743fc32575c56ee8.

* Make the networked object DontDestroyOnLoad

With the game object persistent between scenes noclip doesn't have to
toggle off when changing stages.

The added event to disable noclip when a run ends makes it consistent with
disabling it when disonnecting. This also preserves the old functionality
where noclip would be toggled off after each stage. While within a run it
isn't as helpful, it can be so at the end of it.