Fix vanilla set_scene scene finder hook
The hook isn't supposed to do any name matching, because even if a match
is successful, the scene change is made directly from the argument string
instead of the result returned from the method.

There is also no need to filter for offline/online scenes as the command
itself takes care of that with useful logging messages. Its purpose should
be to just make expansion-locked scenes inaccessible.