Fix dt_bind to not trigger overlapping keybinds, e.g., "x" and "x+y"
This requires a different approach instead of just checking if all the
keys in a combo are currently pressed. An explicit set of modifiers is
definied and each key combo is uniquely identified by its key press and
all its modifiers being held and any remaining modifiers not being held.