Fix the arrow keys not scrolling in the console when time is frozen
This also affects scrolling when time is very slow or too fast. The
solution is to scale the smooth scrolling duration by the current time
scale. For 0 we need to fall back to a special case of checking if the
key has just been pressed else even a brief tap can register for multiple
frames and scroll too much. The key must be pressed each time for one
scroll in this case, but it's better than having no scrolling at all.