Add support for temp items for item commands
Instead of creating separate temp versions of any item commands, a type
argument was added to the pre-existing commands, which leaves the door open
to easily expand this if more item types are added. The positioning of this
parameter could have been at the end so as not to interfere with the
pre-existing command usage, however, it was deemed that this new parameter
would be used more frequently than the target so putting it earlier in the
order was the lesser of two evils.

Updates give_item, remove_item, random_items, remove_item_stacks,
remove_all_items, create_pickup