Modify the autocompletion to use the internal name instead of the index
When selection an autocompletion choice, by the default the first option
was used to substitute that text which in most cases is an index. However,
this can create issues for `create_pickup` as the index cannot
differentiate between distinct items/equipment. Another case is the
interactable catalog, which can be desynced between clients so transmitting
to the server an index can lead to unpredictable results. Both of these
cases can be resolved by using the internal name as the substituted value.